pyproject.toml
[project]
name = "media-mix-modeling"
version = "0.1.0"
description = "Media Mix Modeling with PyMC-Marketing and Databricks"
authors = [
    {name = "Scott McKean", email = "scott.mckean@databricks.com"},
    {name = "Corey Abshire", email = "corey.abshire@databricks.com"},
    {name = "Layla El-Sayed", email = "layla@databricks.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.14"

# Production dependencies (minimal for Databricks deployment)
dependencies = [
    "pymc>=5.16.0",
    "pymc-marketing>=0.12.1",
    "pydantic>=2.11.1",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.6.1",
    "pyyaml>=6.0",
    "arviz>=0.20.0",
    "databricks-connect>=15.0.0",
    "mlflow>=2.20.0",
    "delta-spark>=3.0.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "black>=25.1.0",
    "ruff>=0.8.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "nbformat>=5.10.4",
]

# Agent dependencies (separate from core modeling)
agent = [
    "langgraph>=0.3.22",
    "langchain-core>=0.3.0",
    "databricks-langchain>=0.4.1",
    "openai>=1.70.0",
    "plotly>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "-v",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
